authentication/requirements.txt
python-dotenv
psycopg2
sqlmodel
PyJwt
fastapi[standard]
bcrypt
pydantic
redis>=5.0
asyncpg
sqlalchemy>=1.4.0
asyncio

